Creates new 'play_card_with_trade' which uses the new trade/move checking function for player moves while still maintaining backward compatibility with the AI section (which still uses the old play_card and check_move functions) until a refactor there can be completed.
Pipes play_card API through to all the move information needed for UX design

Moves trade into its own file
